Add PathResolver service and use via DI
