+++ Controllers/PreviewController.cs
@@ -1,6 +1,6 @@
 using Microsoft.AspNetCore.Mvc;
 using Microsoft.AspNetCore.StaticFiles;
-using Microsoft.Extensions.Options;
+using TestProject.Services;
 using System.Globalization;
 using System.IO;
 using System.Text.Json;
@@ -12,18 +12,18 @@ namespace TestProject.Controllers;
 [Route("api/preview")]
 public class PreviewController : ControllerBase
 {
-    private readonly string _root;
+    private readonly PathResolver _pathResolver;
     private readonly FileExtensionContentTypeProvider _contentTypeProvider = new();
 
-    public PreviewController(IOptions<FileExplorerOptions> options)
+    public PreviewController(PathResolver pathResolver)
     {
-        _root = Path.GetFullPath(options.Value.RootPath ?? Directory.GetCurrentDirectory());
+        _pathResolver = pathResolver;
     }
 
     [HttpGet]
     public IActionResult Get([FromQuery] string path)
     {
-        var full = ResolvePath(path);
+        var full = _pathResolver.Resolve(path);
         if (!System.IO.File.Exists(full))
             return NotFound();
 
@@ -39,15 +39,6 @@ public class PreviewController : ControllerBase
             contentType = "application/octet-stream";
 
         return PhysicalFile(full, contentType);
-    }
-
-    private string ResolvePath(string? relative)
-    {
-        relative ??= string.Empty;
-        var combined = Path.GetFullPath(Path.Combine(_root, relative));
-        if (!combined.StartsWith(_root))
-            throw new InvalidOperationException("Invalid path");
-        return combined;
     }
 
     private static string KmlToGeoJson(string kmlContent)
